requirements.txt
fastapi
uvicorn[standard]
google-generativeai
duckduckgo-search
httpx
python-dotenv
pydantic
pydantic-settings
